feat(backend): submission, sheets import, score float + расширенный seed
- score.py: final_score как float для корректной JSON-сериализации
- submission: модель teams.submission JSONB + миграция 0005
- services/submission: heuristic preview и readiness check
- services/sheets_import: импорт команд из Google Sheets через CSV
- api/v1/teams: эндпоинты submission, preview, check, import
- seed.py: 8 команд, 3 жюри, 4 критерия, 4 этапа, реалистичный разброс

+++ backend/app/schemas/score.py
@@ -29,7 +29,7 @@ class ScoreOut(BaseModel):
 class LeaderboardRow(BaseModel):
     team_id: str
     team_name: str
-    final_score: Decimal
+    final_score: float
     judges_count: int
 
 
@@ -37,5 +37,5 @@ class TeamResultOut(BaseModel):
     team_id: str
     team_name: str
     rank: int
-    final_score: Decimal
+    final_score: float
     judges_count: int
